Uploaded df_all2 and df_all3
df_all2 includes the next season's PPR Fantasy Points Allowed for each team but is slightly bugged. Right now, it uses the next row's PPR Fantasy Points Allowed as its current metric for Next Season's PPR Fantasy Points Allowed, but this poses an issue when the next row is not the next season. This happens when the data switches teams, which would not be a problem if they all ended at 2019, but as there are various name changes, this gets a bit messy.

df_all3 uses data only from 2000 onward and fills in each team's 2019 Next Season's PPR Fantasy Points Allowed as NA. This should be usable for modeling.